Move time setting checkpoint to main.dart
